# etc/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "SATI - BackOffice",
    "version": "v1"
  },
  "paths": {
    "/api/BOArchivo": {
      "get": {
        "tags": [
          "BOArchivo"
        ],
        "operationId": "GetArchivos",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "Todo": {
                    "type": "boolean"
                  },
                  "Id": { },
                  "IdRef": {
                    "type": "integer",
                    "format": "int32"
                  },
                  "IdG": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "Date": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "Search": {
                    "type": "string"
                  },
                  "PageSize": {
                    "type": "integer",
                    "format": "int32"
                  },
                  "PageNumber": {
                    "type": "integer",
                    "format": "int32"
                  },
                  "Sort": {
                    "type": "string"
                  },
                  "SortDir": {
                    "type": "string"
                  }
                }
              },
              "encoding": {
                "Todo": {
                  "style": "form"
                },
                "Id": {
                  "style": "form"
                },
                "IdRef": {
                  "style": "form"
                },
                "IdG": {
                  "style": "form"
                },
                "Date": {
                  "style": "form"
                },
                "Search": {
                  "style": "form"
                },
                "PageSize": {
                  "style": "form"
                },
                "PageNumber": {
                  "style": "form"
                },
                "Sort": {
                  "style": "form"
                },
                "SortDir": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EArchivoIEnumerableApiResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "BOArchivo"
        ],
        "parameters": [
          {
            "name": "esTemporal",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/EArchivo"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EArchivo"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/EArchivo"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/EArchivo"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "BOArchivo"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/BOArchivo/{id}": {
      "get": {
        "tags": [
          "BOArchivo"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "BOArchivo"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/EArchivo"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EArchivo"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/EArchivo"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/EArchivo"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EArchivo": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "codigoSistema": {
            "type": "string",
            "nullable": true
          },
          "claveRelacion": {
            "type": "string",
            "nullable": true
          },
          "tipoEDocumentoId": {
            "type": "integer",
            "format": "int32"
          },
          "estadoEDocumentoId": {
            "type": "integer",
            "format": "int32"
          },
          "requerimientoId": {
            "type": "integer",
            "format": "int32"
          },
          "solicitudId": {
            "type": "string",
            "nullable": true
          },
          "tipoSolicitudId": {
            "type": "string",
            "nullable": true
          },
          "cuit": {
            "type": "string",
            "nullable": true
          },
          "nombreArchivo": {
            "type": "string",
            "nullable": true
          },
          "descripcion": {
            "type": "string",
            "nullable": true
          },
          "ruta": {
            "type": "string",
            "nullable": true
          },
          "fechaIndexacion": {
            "type": "string",
            "format": "date-time"
          },
          "hash": {
            "type": "string",
            "nullable": true
          },
          "archivo": {
            "nullable": true
          },
          "claveTemporal": {
            "type": "integer",
            "format": "int64"
          }
        },
        "additionalProperties": false
      },
      "EArchivoIEnumerableApiResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EArchivo"
            },
            "nullable": true
          },
          "meta": {
            "$ref": "#/components/schemas/Metadata"
          }
        },
        "additionalProperties": false
      },
      "Metadata": {
        "type": "object",
        "properties": {
          "totalCount": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "currentPage": {
            "type": "integer",
            "format": "int32"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "hasNextPage": {
            "type": "boolean"
          },
          "hasPreviousPage": {
            "type": "boolean"
          },
          "nextPageUrl": {
            "type": "string",
            "nullable": true
          },
          "previousPageUrl": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
    }
  }
}